requirements/local.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/local.in
#
asgiref==3.4.1
    # via django
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
django==3.2.9
    # via
    #   -r requirements/local.in
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.12.4
    # via
    #   -r requirements/local.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.0.0
    # via -r requirements/local.in
idna==3.3
    # via requests
pyjwt==2.3.0
    # via djangorestframework-simplejwt
pytz==2021.3
    # via django
requests==2.26.0
    # via -r requirements/local.in
sqlparse==0.4.2
    # via django
urllib3==1.26.7
    # via requests
